test(chat): fold subscribe-effect-timing regression into chat-adapter.test.ts
The standalone `chat-adapter-sync-write-fixture.svelte` + its own test
file tanked the package's Svelte function-coverage ratchet (82.83% ->
~70%, well under the 81.34% baseline) because it mounted a fresh,
minimally-exercised copy of Chat's whole subcomponent tree — the same
class of problem `isTransientTestArtifact` already calls out for SSR
probes ("exercise only a fraction of each component's functions").

Verified locally on independent clean worktrees off both origin/main
and this branch: removing the standalone fixture restores the ratchet
to passing (confirmed via `bun run --filter=@lostgradient/chat
test:coverage`), and Copilot flagged the file's ResizeObserver/
IntersectionObserver global overrides never being restored, unlike
sibling test files in this suite.

Move the regression test into the existing `chat-adapter.test.ts`
"subscribe lifecycle" block, reusing its `mountChat` helper (which
every other test in the file already exercises, so no new low-coverage
render tree is introduced), and replace the `.svelte` fixture with a
tiny `subscribe-event-log.svelte.ts` reactive class — real `$state`
without a component. No behavior or documentation change; same
assertions as before (deferring the `subscribe` write with
`queueMicrotask` still lands after composer input, without throwing).
